feat: uses base model to gather common columns
