+++ src/main/java/com/casino_gacha/webapp/models/Card.java
@@ -1,11 +1,7 @@
 package com.casino_gacha.webapp.models;
 
-import java.sql.Timestamp;
 import java.util.List;
 import java.util.UUID;
-
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
 
 import com.casino_gacha.webapp.enums.CardGender;
 
@@ -21,7 +17,7 @@ import jakarta.persistence.Table;
 
 @Entity
 @Table(name = "card")
-public class Card {
+public class Card extends Base {
 
   @Id
   @GeneratedValue(strategy = GenerationType.AUTO)
@@ -46,26 +42,6 @@ public class Card {
 
   @OneToMany(mappedBy = "id")
   private List<Image> images;
-
-  @CreatedDate
-  @Column(name = "created_at", nullable = false)
-  private Timestamp createdAt;
-
-  @Column(name = "created_by")
-  private UUID createdBy;
-
-  @LastModifiedDate
-  @Column(name = "updated_at", nullable = false)
-  private Timestamp updatedAt;
-
-  @Column(name = "updated_by")
-  private UUID updatedBy;
-
-  @Column(name = "deleted_at")
-  private Timestamp deletedAt;
-
-  @Column(name = "deleted_by")
-  private UUID deletedBy;
 
   public void setId(UUID id) {
     this.id = id;
@@ -113,53 +89,5 @@ public class Card {
 
   public int getValue() {
     return value;
-  }
-
-  public void setCreatedAt(Timestamp createdAt) {
-    this.createdAt = createdAt;
-  }
-
-  public Timestamp getCreatedAt() {
-    return createdAt;
-  }
-
-  public void setCreatedBy(UUID createdBy) {
-    this.createdBy = createdBy;
-  }
-
-  public UUID getCreatedBy() {
-    return createdBy;
-  }
-
-  public void setUpdatedAt(Timestamp updatedAt) {
-    this.updatedAt = updatedAt;
-  }
-
-  public Timestamp getUpdatedAt() {
-    return updatedAt;
-  }
-
-  public void setUpdatedBy(UUID updatedBy) {
-    this.updatedBy = updatedBy;
-  }
-
-  public UUID getUpdatedBy() {
-    return updatedBy;
-  }
-
-  public void setDeletedAt(Timestamp deletedAt) {
-    this.deletedAt = deletedAt;
-  }
-
-  public Timestamp getDeletedAt() {
-    return deletedAt;
-  }
-
-  public void setDeletedBy(UUID deletedBy) {
-    this.deletedBy = deletedBy;
-  }
-
-  public UUID getDeletedBy() {
-    return deletedBy;
   }
 }
