+++ src/main/java/com/casino_gacha/webapp/models/CasinoUser.java
@@ -1,11 +1,5 @@
 package com.casino_gacha.webapp.models;
 
-import java.sql.Timestamp;
-import java.util.UUID;
-
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-
 import com.casino_gacha.webapp.models.coumpound_keys.CasinoUserPK;
 
 import jakarta.persistence.Column;
@@ -15,7 +9,7 @@ import jakarta.persistence.Table;
 
 @Entity
 @Table(name = "casino_user")
-public class CasinoUser {
+public class CasinoUser extends Base {
 
   @EmbeddedId
   private CasinoUserPK id;
@@ -25,26 +19,6 @@ public class CasinoUser {
 
   @Column(name = "wallet", nullable = false, columnDefinition = "int default 0")
   private int wallet;
-
-  @CreatedDate
-  @Column(name = "created_at", nullable = false)
-  private Timestamp createdAt;
-
-  @Column(name = "created_by")
-  private UUID createdBy;
-
-  @LastModifiedDate
-  @Column(name = "updated_at", nullable = false)
-  private Timestamp updatedAt;
-
-  @Column(name = "updated_by")
-  private UUID updatedBy;
-
-  @Column(name = "deleted_at")
-  private Timestamp deletedAt;
-
-  @Column(name = "deleted_by")
-  private UUID deletedBy;
 
   public void setNickname(String nickname) {
     this.nickname = nickname;
@@ -60,53 +34,5 @@ public class CasinoUser {
 
   public int getWallet() {
     return wallet;
-  }
-
-  public void setCreatedAt(Timestamp createdAt) {
-    this.createdAt = createdAt;
-  }
-
-  public Timestamp getCreatedAt() {
-    return createdAt;
-  }
-
-  public void setCreatedBy(UUID createdBy) {
-    this.createdBy = createdBy;
-  }
-
-  public UUID getCreatedBy() {
-    return createdBy;
-  }
-
-  public void setUpdatedAt(Timestamp updatedAt) {
-    this.updatedAt = updatedAt;
-  }
-
-  public Timestamp getUpdatedAt() {
-    return updatedAt;
-  }
-
-  public void setUpdatedBy(UUID updatedBy) {
-    this.updatedBy = updatedBy;
-  }
-
-  public UUID getUpdatedBy() {
-    return updatedBy;
-  }
-
-  public void setDeletedAt(Timestamp deletedAt) {
-    this.deletedAt = deletedAt;
-  }
-
-  public Timestamp getDeletedAt() {
-    return deletedAt;
-  }
-
-  public void setDeletedBy(UUID deletedBy) {
-    this.deletedBy = deletedBy;
-  }
-
-  public UUID getDeletedBy() {
-    return deletedBy;
   }
 }
